+++ blocks/load.js
@@ -2,5 +2,5 @@ import './donation-form/index';
 import './donation-form-grid/index';
 import './donor-wall/index';
 
-// Import milestone block from Milestones src
-import '../src/Milestones/resources/js/index';
+// Import total block from Totals src
+import '../src/Totals/resources/js/index';