+++ give.php
@@ -46,7 +46,7 @@ use Give\Controller\Form as FormRouteController;
 use Give\ServiceProviders\LegacyServiceProvider;
 use Give\ServiceProviders\RestAPI;
 use Give\ServiceProviders\Onboarding;
-use Give\ServiceProviders\Milestones;
+use Give\ServiceProviders\Totals;
 use Give\ServiceProviders\ServiceProvider;
 
 // Exit if accessed directly.
@@ -134,7 +134,7 @@ final class Give {
 		LegacyServiceProvider::class,
 		RestAPI::class,
 		Onboarding::class,
-		Milestones::class,
+		Totals::class,
 	];
 
 	/**