+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Milestone now supports aggregating different metrics (Revenue, Donations, Donors) (#5244)
 -   Milestone title and description now support {} tags (#5242)
 -   Milestone now supports a custom Call To Action URL and text (#5262)
+-   Milestone block has been replaced by the Totals block, with features to match the Give Totals shortcode (#5264)
 
 ## [2.8.0] - 2020-08-31